Update the Authorization

+++ src/AppDiv.SmartAgency.Application/Features/LookUps/Command/Delete/DeleteLookUpCommand.cs
@@ -7,7 +7,7 @@ using MediatR;
 
 namespace AppDiv.SmartAgency.Application.Features.LookUps.Command.Delete
 {
-    public class DeleteLookUpCommand: IRequest<String>
+    public class DeleteLookUpCommand : IRequest<String>
     {
         public Guid Id { get; private set; }
 
@@ -17,14 +17,14 @@ namespace AppDiv.SmartAgency.Application.Features.LookUps.Command.Delete
         }
     }
 
-   
+
     // lookUp delete command handler with string response as output
     public class DeleteLookUpCommmandHandler : IRequestHandler<DeleteLookUpCommand, String>
     {
         private readonly ILookUpRepository _lookUpRepository;
         public DeleteLookUpCommmandHandler(ILookUpRepository lookUpRepository)
         {
-            _lookUpRepository= lookUpRepository;
+            _lookUpRepository = lookUpRepository;
         }
 
         public async Task<string> Handle(DeleteLookUpCommand request, CancellationToken cancellationToken)
@@ -32,18 +32,22 @@ namespace AppDiv.SmartAgency.Application.Features.LookUps.Command.Delete
             try
             {
                 var lookUpEntity = await _lookUpRepository.GetByIdAsync(request.Id);
+                if (lookUpEntity is not null)
+                {
+                    await _lookUpRepository.DeleteAsync(lookUpEntity.Id);
+                    await _lookUpRepository.SaveChangesAsync(cancellationToken);
+                }
+                else return $"No entity found with an Id {request.Id}";
 
-                await _lookUpRepository.DeleteAsync(lookUpEntity.Id);
-                await _lookUpRepository.SaveChangesAsync(cancellationToken);
             }
             catch (Exception exp)
             {
-                throw (new ApplicationException(exp.Message));
+                throw new ApplicationException(exp.Message);
             }
 
             return "LookUp information has been deleted!";
         }
     }
 }
- 
+
